cloud/ingest_locations/requirements.txt
asttokens==2.4.1
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.8
cloudevents==1.11.0
comm==0.2.2
db-dtypes==1.3.0
debugpy==1.8.7
decorator==5.1.1
deprecation==2.1.0
executing==2.1.0
Flask==3.1.0
functions-framework==3.8.2
google-api-core==2.21.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.26.0
google-cloud-core==2.4.1
google-cloud-secret-manager==2.20.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.2
grpcio-status==1.66.2
gunicorn==23.0.0
idna==3.10
ipykernel==6.29.5
ipython==8.28.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.5
jupyter_client==8.6.3
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
numpy==2.1.2
oauthlib==3.2.2
packaging==24.1
pandas==2.2.3
pandas-gbq==0.24.0
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
proto-plus==1.24.0
protobuf==5.28.2
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydata-google-auth==1.8.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyzmq==26.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
setuptools==75.1.0
six==1.16.0
stack-data==0.6.3
tornado==6.4.1
traitlets==5.14.3
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
